requirements.txt
langchain-groq
langchain-community
python-dotenv
uvicorn
fastapi
pydantic
streamlit
langgraph
langchain-tavily